backend/pyproject.toml
[project]
name = "courseforge-backend"
version = "0.1.0"
description = "CourseForge backend API and pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "arq>=0.26.1",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "anthropic>=0.39.0",
    "openai>=1.56.0",
    "python-docx>=1.1.2",
    "pdf2image>=1.17.0",
    "trafilatura>=1.12.0",
    "playwright>=1.49.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.18",
    "structlog>=24.4.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
    "aiosqlite>=0.20.0",
    "respx>=0.22.0",
]
